requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo[srv]==4.5.0
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
pydantic==2.5.0